requirements.txt
aiogram==2.23.1
aiohttp==3.8.3
aioredis==2.0.1
APScheduler==3.8.1
pandas==1.5.2
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
python-i18n==0.3.9
pytz==2022.6
redis==4.4.0
requests==2.28.1
# hiredis==2.0.0
tzlocal<3.0